Delete First_NN_M5_Simple_Fashion_MNIST.ipynb
This file had verbose set to 1 on the evaluate methods which generated thousands of "=" characters and prevented it from opening in the github website.  Since it would not open in the github website users could not click the "run in Colab" link.  The file M5_First_Fashion_MNIST.ipynb sets verbose to 0, does not have issues opening, and therefore replaces this file.